requirements.txt
spotipy
flask
pandas
flask_session
PyJWT
pyrebase4